[UPDATE] Move all component logic from App() to hooks in order to delegate each beahavior. However, this caused in misfunctionnement in autoupdate => to fix
